little fix

+++ src/main/java/com/dragonfix/mixin/mixins/opencomputers/TextBufferRenderCacheMixin.java
@@ -37,7 +37,7 @@ public abstract class TextBufferRenderCacheMixin {
 
         Tuple2<Object, Object> viewport = buffer.viewport();
         renderer()
-            .drawBuffer(buffer.data(), ((Integer) viewport._1()).intValue(), ((Integer) viewport._2()).intValue());
+            .drawBuffer(buffer.data(), (Integer) viewport._1(), (Integer) viewport._2());
 
         RenderState.checkError(getClass().getName() + ".render: leaving");
         ci.cancel();
